Simplify sif network into igraph network graph object

This function removes duplicated edges and loops to create an igraph graph object from tab delimited sif formatted network file.

Usage

networkSimplify(sifNetwork, directed = FALSE)

Arguments

sifNetwork

A file with sif network format (There are three columns in the file separated by tab, nodeA interactionType nodeB )

directed

Logical, treat network as directed or undirected graph

Value

a igraph graph object

Details

For undirected graph, networkSimplify removes duplicated edges and loops to create an igraph graph object from tab delimited sif formatted network file.

For directed graph, networkSimplify selects the first edge and removes the rest duplicated edges and loops to create an igraph graph object from tab delimited sif formatted network file.

Author

Eric Minwei Liu, emliu.research@gmail.com

Examples

data(netbox2010)

sifNetwork <- netbox2010$network
graphReduced <- networkSimplify(sifNetwork, directed = FALSE)
#> Loading network of 9264 nodes and 157780 interactions
#> Treated as undirected network 
#> Removing multiple interactions and loops
#> Returning network of 9264 nodes and 68111 interactions